api/requirements.txt
coreapi==2.3.3
Django==3.1.1
django-allauth==0.44.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
drf-yasg==1.17.1
PyYAML==5.3.1